resolved: add flags to DnsResourceKey
mDNS defines the highest bit in an answer record (0x8000) as
'flush cache' bit (RFC 6762, 10.2). In order to support that,
introduce a flags field in DnsResourceKey, and set the
DNS_RESOURCE_KEY_CACHE_FLUSH from dns_packet_read_key() in case
the bit is set in an answer.

The bit is not looked at by any component yet.